server/requirements.txt
flask==3.0.3
flask-socketio==5.3.6
flask-cors==4.0.1
pymongo==4.7.2
python-dotenv==1.0.1
pdfplumber==0.11.1
requests==2.32.3
werkzeug==3.0.3
eventlet==0.36.1
openai>=1.0.0
